resolved slowly ds

+++ rd-api/src/main/java/it/clevercom/echo/rd/repository/IService_rd_Repository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.clevercom.echo.rd.model.entity.Service;
 
-public interface IService_rd_Repository extends JpaRepository<Service, Integer> {
+public interface IService_rd_Repository extends JpaRepository<Service, Long> {
 
 }
